requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
openai>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.6
